% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsaves.R
\name{ggisaves}
\alias{ggisaves}
\title{Save interactive plot to common filetypes}
\usage{
ggisaves(
  plot,
  prefix,
  outfolder,
  title = prefix,
  knitrOptions,
  background = "white",
  width = 8,
  height = 5,
  dpi = 300
)
}
\arguments{
\item{prefix}{filename without extension}

\item{outfolder}{folder to save files in}

\item{title}{Text to use as the title of the generated page.}

\item{knitrOptions}{A list of \pkg{knitr} chunk options.}

\item{background}{Text string giving the html background color of the widget.
Defaults to white.}

\item{width, height}{Plot size in inches. If not supplied, uses the size of current graphics device.}
}
\value{
Run for its side effects. Returns invisible(NULL)
}
\description{
Save interactive plot to common filetypes
}
\examples{

\dontrun{
 library(ggiraph)
 library(ggplot2)

 gg_plot <- iris |>
     ggplot(aes(Sepal.Length, Sepal.Width)) +
     geom_point_interactive(aes(data_id  = Species, tooltip = Species))

   interactive_gg_plot <- girafe(ggobj = gg_plot)

   ggisaves(interactive_gg_plot, prefix = "my_interactive_plot", outfolder = ".")
}
}
