% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsaves.R
\name{ggsaves}
\alias{ggsaves}
\title{Save plot to common filetypes}
\usage{
ggsaves(plot, prefix, outfolder, dpi = 300, width = 8, height = 5, ...)
}
\arguments{
\item{plot}{Plot to save, defaults to last plot displayed.}

\item{prefix}{filename without extension}

\item{outfolder}{folder to save files in}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}

\item{width, height}{Plot size in units ("in", "cm", "mm", or "px"). If not supplied, uses the size of current graphics device.}

\item{...}{
  Arguments passed on to \code{\link[ggplot2:ggsave]{ggplot2::ggsave}}
  \describe{
    \item{\code{filename}}{File name to create on disk.}
    \item{\code{device}}{Device to use. Can either be a device function
(e.g. \link{png}), or one of "eps", "ps", "tex" (pictex),
"pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only). If
\code{NULL} (default), the device is guessed based on the \code{filename} extension.}
    \item{\code{path}}{Path of the directory to save plot to: \code{path} and \code{filename}
are combined to create the fully qualified file name. Defaults to the
working directory.}
    \item{\code{scale}}{Multiplicative scaling factor.}
    \item{\code{width,height}}{Plot size in units expressed by the \code{units} argument.
If not supplied, uses the size of the current graphics device.}
    \item{\code{units}}{One of the following units in which the \code{width} and \code{height}
arguments are expressed: \code{"in"}, \code{"cm"}, \code{"mm"} or \code{"px"}.}
    \item{\code{limitsize}}{When \code{TRUE} (the default), \code{ggsave()} will not
save images larger than 50x50 inches, to prevent the common error of
specifying dimensions in pixels.}
    \item{\code{bg}}{Background colour. If \code{NULL}, uses the \code{plot.background} fill value
from the plot theme.}
    \item{\code{create.dir}}{Whether to create new directories if a non-existing
directory is specified in the \code{filename} or \code{path} (\code{TRUE}) or return an
error (\code{FALSE}, default). If \code{FALSE} and run in an interactive session,
a prompt will appear asking to create a new directory when necessary.}
  }}
}
\value{
Run for its side effects. Returns invisible(NULL)
}
\description{
Save plot to common filetypes
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(ggplot2)

# Create Plot
gg_plot <- ggplot(mtcars, aes(mpg, cyl)) + geom_point()

# Save as all common file formats
ggsaves(gg_plot, prefix = "myplot",  outfolder = "plots")
\dontshow{\}) # examplesIf}
}
