% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsaves.R
\name{ggisaves}
\alias{ggisaves}
\title{Save interactive plot to common filetypes}
\usage{
ggisaves(
  plot,
  prefix,
  outfolder,
  title = prefix,
  knitrOptions,
  background = "white",
  dpi = 300,
  width = 8,
  height = 5
)
}
\arguments{
\item{plot}{a htmlwidget plot to save in standard file formats.}

\item{prefix}{filename without extension}

\item{outfolder}{folder to save files in}

\item{title}{Text to use as the title of the generated page.}

\item{knitrOptions}{A list of \pkg{knitr} chunk options.}

\item{background}{Text string giving the html background color of the widget.
Defaults to white.}

\item{dpi}{resolution to save raster image files.}

\item{width, height}{Plot size in inches.}
}
\value{
Run for its side effects. Returns invisible(NULL)
}
\description{
Save interactive plot to common filetypes
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(ggplot2)
library(ggiraph)
# Make static ggplot
gg_plot <- ggplot(iris, aes(Sepal.Length, Sepal.Width)) +
  geom_point_interactive(aes(data_id  = Species, tooltip = Species))

# Create interactive HTMLwidget
interactive_gg_plot <- girafe(ggobj = gg_plot)

# Save HTMLwidget
ggisaves(interactive_gg_plot, prefix = "my_interactive_plot", outfolder = "interactive_plots")
\dontshow{\}) # examplesIf}
}
